Completed Part 11 of Tutorial
- Modified app structure to use flask Blueprint and application factory
pattern
- Created Config object
- create_app function is used to create an instant of the app
