+++ flaskblog/models.py
@@ -1,4 +1,5 @@
-from flaskblog import db, login_manager, app
+from flask import current_app
+from flaskblog import db, login_manager
 from datetime import datetime
 from flask_login import UserMixin
 from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
@@ -25,12 +26,12 @@ class User(db.Model, UserMixin):
         return f'User({self.username}","{self.email}", "{self.image_file}")'
 
     def get_reset_token(self, expires_secs=1800):
-        s = Serializer(app.config['SECRET_KEY'], expires_secs)
+        s = Serializer(current_app.config['SECRET_KEY'], expires_secs)
         return s.dumps({'user_id': self.id}).decode('utf-8')
     
     @staticmethod
     def verify_reset_token(token):
-        s = Serializer(app.config['SECRET_KEY'])
+        s = Serializer(current_app.config['SECRET_KEY'])
         try:
             user_id = s.loads(token)['user_id']
         except:
