+++ flaskblog/templates/author_posts.html
@@ -6,10 +6,10 @@
               <img class="rounded-circle article-img" src="{{ url_for('static', filename='profile_pictures/' + post.author.image_file) }}" />
                 <div class="media-body">
                   <div class="article-metadata">
-                    <a class="mr-2" href="{{ url_for('author_posts', username=post.author.username) }}">{{ post.author.username }}</a>
+                    <a class="mr-2" href="{{ url_for('users.author_posts', username=post.author.username) }}">{{ post.author.username }}</a>
                     <small class="text-muted">{{ post.date_posted.strftime('%d %B %Y') }}</small>
                   </div>
-                  <h2><a class="article-title" href="{{ url_for('post', post_id=post.id) }}">{{ post.title }}</a></h2>
+                  <h2><a class="article-title" href="{{ url_for('posts.post', post_id=post.id) }}">{{ post.title }}</a></h2>
                   <p class="article-content">{{ post.content }}</p>
                 </div>
               </article>
@@ -17,9 +17,9 @@
         {% for page_num in posts.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
           {% if page_num %}
             {% if page_num == posts.page %}
-              <a href="{{url_for('author_posts', page=page_num, username=author.username)}}" class="btn btn-info mb-4">{{ page_num }}</a>              
+              <a href="{{url_for('users.author_posts', page=page_num, username=author.username)}}" class="btn btn-info mb-4">{{ page_num }}</a>              
             {% else %}
-              <a href="{{url_for('author_posts', page=page_num, username=author.username)}}" class="btn btn-outline-info mb-4">{{ page_num }}</a>
+              <a href="{{url_for('users.author_posts', page=page_num, username=author.username)}}" class="btn btn-outline-info mb-4">{{ page_num }}</a>
             {% endif %}
           {% else %}
           ...
