+++ flaskblog/templates/layout.html
@@ -26,18 +26,18 @@
                     <div class="collapse navbar-collapse" id="navbarToggle">
                         <!-- Navbar left Side -->
                         <div class="navbar-nav mr-auto">
-                            <a class="nav-item nav-link" href={{ url_for('home') }}>Home</a>
-                            <a class="nav-item nav-link" href={{ url_for('about') }}>About</a>
+                            <a class="nav-item nav-link" href={{ url_for('main.home') }}>Home</a>
+                            <a class="nav-item nav-link" href={{ url_for('main.about') }}>About</a>
                         </div>
                         <!-- Navbar Right Side -->
                         <div class="navbar-nav">
                           {% if current_user.is_authenticated %}
-                            <a class="nav-item nav-link" href={{ url_for('new_post') }}>New Post</a>
-                            <a class="nav-item nav-link" href={{ url_for('account') }}>Account</a>
-                            <a class="nav-item nav-link" href={{ url_for('logout') }}>Logout</a>
+                            <a class="nav-item nav-link" href={{ url_for('posts.new_post') }}>New Post</a>
+                            <a class="nav-item nav-link" href={{ url_for('users.account') }}>Account</a>
+                            <a class="nav-item nav-link" href={{ url_for('users.logout') }}>Logout</a>
                           {% else %}
-                            <a class="nav-item nav-link" href={{ url_for('login') }}>Login</a>
-                            <a class="nav-item nav-link" href={{ url_for('register') }}>Register</a>
+                            <a class="nav-item nav-link" href={{ url_for('users.login') }}>Login</a>
+                            <a class="nav-item nav-link" href={{ url_for('users.register') }}>Register</a>
                           {% endif %}
                         </div>
                     </div>
