+++ flaskblog/templates/login.html
@@ -40,14 +40,14 @@
             <div class="form-group">
                 {{ form.submit(class="btn btn-outline-info") }}
                 <small class="text-muted ml-2">
-                    <a href="{{ url_for('reset_request') }}">Forgot password?</a>
+                    <a href="{{ url_for('users.reset_request') }}">Forgot password?</a>
                 </small>
             </div>
         </form>
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Need an account? <a class="ml-2" href={{ url_for('register') }}>Sign Up</a>
+            Need an account? <a class="ml-2" href={{ url_for('users.register') }}>Sign Up</a>
         </small>
         
     </div>
