+++ flaskblog/templates/register.html
@@ -67,7 +67,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Already have an account? <a class="ml-2" href={{ url_for('login') }}>Sign In</a>
+            Already have an account? <a class="ml-2" href={{ url_for('users.login') }}>Sign In</a>
         </small>
     </div>
     {% endblock content %}
